pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["siggu <m0_w_0m@naver.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"
pyjwt = "^2.8.0"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
